Add file explorer
In neovim the current file explorer is implemented with the coc-explorer
module.
For a more lightweight (and external plugin-free configuration) the
configuration for netrw have been added as well to the configuration but
commented out.

The minimal vim configuration uses just netrw (same configuration of
neovim).

In the future I want to add custom keybindings for standard operations
on files (copy, paste, delete, create directory, etc.) and test the
netrw config versus the coc-explorer.