Add clever-f plugin
clever-f plugin allows to optimize and enhance the navigation using
f/F/t/T Vim commands allowing:
* search through multiple lines
* avoid , and ; keybindings but using only f(or t) and F(or T) for
forward and backward search. In this way the search is optimized and ','
 ';' keys are free for other bindings.
* the user can choose the granularity of the navigation.

The options I choose involve navigation with matching-case
(with smart-case option turned on) and looking for the pattern in
multiple lines.

In this state the plugin is 'turned off' with any other Vim motion (e.g.
h-j-k-l). In the future I would like to add a key binding (e.g.
<Esc><Esc>) for reset the navigation.