Add tagbar plugin
Implementation of tagbar using vista.vim plugin.
The plugin depends on ctags to generate tags and the sidebar is toggled
on the right of the screen.

In conduction with CoC-Explorer I can work with the file explorer and
the tagbar for an IDE-like setup.

Other changes regard the installation of a plugin that fixes the
CursorHold and CursorHoldI performances bugs on NeoVim.