Change fix_key_direction option
Explicitly changed this option. In this way the direction of the search
can be set by the first command (f/F/t/T) and then pressing just (f or
t) we can navigate along the chosen direction (lower-case=forward,
upper-case=backward).

With this set-up the plugin in more efficient and in the case of
backward navigation there's no need for pressing capital letters.