Change text file formatting
Custom format option for LaTeX were not working with LaTeX's custom
enviorments (equations, figures, tables, etc.) messing all the
indentation.
To solve the problem the wrapping (with broken lines) is removed and
linebreaks option is added to the configuration that allows Vim to
wrap only words that does not fit on the screen.

These set of instructions is used for general text files and thus added
to both LaTeX and Markdown filetypes.